# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Gestión de Pedidos de Aberturas",
        "description": "API REST para gestionar los pedidos de una empresa de aberturas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/users/{id}/permissions": {
            "get": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Ver los permisos de un usuario",
                "description": "Devuelve los permisos individuales de un usuario operativo (cliente o vendedor). Solo administradores.",
                "operationId": "a28653de000e844fcdfc18f25d08228d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permisos obtenidos correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    },
                    "403": {
                        "description": "No tenés permiso para realizar esta acción."
                    },
                    "404": {
                        "description": "Usuario no encontrado."
                    },
                    "422": {
                        "description": "El usuario no es un usuario operativo."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Cambiar los permisos de un usuario",
                "description": "Activa o desactiva permisos de un usuario operativo. Solo se cambian los permisos enviados. Solo administradores.",
                "operationId": "1c9ce825a423c3c15ae646cd1e08106f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 2
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "create_orders": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "view_own_orders": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "view_all_orders": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "verify_payments": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "update_item_status": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "edit_ship_date": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "register_balance_payment": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permisos actualizados correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    },
                    "403": {
                        "description": "No tenés permiso para realizar esta acción."
                    },
                    "404": {
                        "description": "Usuario no encontrado."
                    },
                    "422": {
                        "description": "Los datos proporcionados no son válidos."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación generado por Laravel Sanctum.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}